# contracts/openapi/core.yaml
openapi: 3.1.0
info:
  title: mmdash Core API
  version: 0.1.0
  description: |
    Stable HTTP contract exposed by the Go Core Server. Browser clients use it
    through Web BFF, and Agent clients use it through MCP Gateway.
servers:
  - url: http://localhost:8080
    description: Local Core Server
...
paths:
...
  /v1/projects:
    get:
      tags: [Projects]
      summary: List projects visible to the current identity
      operationId: projects.list
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Visible collaborative projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      tags: [Projects]
      summary: Create a collaborative project
      operationId: projects.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created with the caller as owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /v1/projects/{projectId}/repository/sync:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Repo]
      summary: Coalesce a manual synchronization request
      operationId: repo.sync.request
      responses:
        "202":
          description: Synchronization request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
...
  /v1/projects/{projectId}/models:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Model]
      summary: Read the project Model overview
      operationId: model.get
      responses:
        "200":
          description: The single source, discovered pages, questions, and current synchronization state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelOverview"
  /v1/projects/{projectId}/models/source:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Model]
      summary: Read the single project Model source
      operationId: model.source.get
      responses:
        "200":
          description: Project Model source and automatic synchronization countdown.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSource"
  /v1/projects/{projectId}/models/source/sync:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Model]
      summary: Rediscover the configured Notion root and its descendant pages
      operationId: model.source.sync
      responses:
        "202":
          description: Source discovery synchronization accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSync"
  /v1/projects/{projectId}/models/questions:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Model]
      summary: List Model questions
      operationId: model.questions.list
      responses:
        "200":
          description: Questions ordered by their explicit position.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelQuestionList"
    post:
      tags: [Model]
      summary: Create a Model question bound to one discovered Notion child page
      operationId: model.questions.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelQuestionRequest"
      responses:
        "201":
          description: Created Model question.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelQuestion"
  /v1/projects/{projectId}/models/questions/{questionId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelQuestionId"
    get:
      tags: [Model]
      summary: Read one Model question
      operationId: model.questions.get
      responses:
        "200":
          description: Question and its current immutable Snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelQuestionDetail"
    patch:
      tags: [Model]
      summary: Update Model question display metadata or bound descendant page
      operationId: model.questions.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelQuestionRequest"
      responses:
        "200":
          description: Updated Model question.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelQuestion"
    delete:
      tags: [Model]
      summary: Remove a Model question while retaining its referenced Artifacts
      operationId: model.questions.delete
      responses:
        "204": { description: Model question removed. }
  /v1/projects/{projectId}/models/questions/{questionId}/sync:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelQuestionId"
    post:
      tags: [Model]
      summary: Manually synchronize one Model question
      description: A manual synchronization resets the project automatic synchronization countdown.
      operationId: model.questions.sync
      responses:
        "202":
          description: Question synchronization accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSync"
  /v1/projects/{projectId}/models/questions/{questionId}/snapshots:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelQuestionId"
    get:
      tags: [Model]
      summary: List immutable Snapshots for one Model question
      operationId: model.snapshots.list
      responses:
        "200":
          description: Newest-first Snapshot timeline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSnapshotList"
  /v1/projects/{projectId}/models/questions/{questionId}/snapshots/{snapshotId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelQuestionId"
      - $ref: "#/components/parameters/ModelSnapshotId"
    get:
      tags: [Model]
      summary: Read one fixed Model Snapshot
      operationId: model.snapshots.get
      responses:
        "200":
          description: Immutable content plus editable tags and version note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSnapshot"
    patch:
      tags: [Model]
      summary: Update user-managed Snapshot tags and version note
      operationId: model.snapshots.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelSnapshotRequest"
      responses:
        "200":
          description: Updated Snapshot metadata; immutable content and Hash are unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSnapshot"
  /v1/projects/{projectId}/models/questions/{questionId}/diff:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelQuestionId"
      - name: from_snapshot_id
        in: query
        required: true
        schema: { type: string, format: uuid }
      - name: to_snapshot_id
        in: query
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Model]
      summary: Compare two Model Snapshots at character granularity
      operationId: model.snapshots.diff
      responses:
        "200":
          description: Block-preserving diff with contiguous character runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelDiff"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or opaque token
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ModelQuestionId:
      name: questionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ModelSnapshotId:
      name: snapshotId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ProjectRole:
      type: string
      enum: [owner, maintainer, editor, viewer, agent, box]
    Project:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - problem_title
        - problem_summary
        - project_constraints
        - source_artifact_ids
        - created_by
        - created_at
        - updated_at
        - role
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        problem_title:
          type: string
        problem_summary:
          type: string
        project_constraints:
          type: array
          items:
            type: string
        source_artifact_ids:
          type: array
          items:
            type: string
        created_by:
          type: string
          format: uuid
        archived_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          description: Time the owner moved the project to the recycle bin.
        purge_at:
          type: string
          format: date-time
          description: Time after which the project can no longer be restored.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        role:
          $ref: "#/components/schemas/ProjectRole"
    ProjectList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        problem_title:
          type: string
        problem_summary:
          type: string
        project_constraints:
          type: array
          items:
            type: string
        source_artifact_ids:
          type: array
          items:
            type: string
...
    RepoProvider:
      type: string
      enum: [github, local]
    RepoWorkspaceKind:
      type: string
      enum: [code, article, result]
    RepositoryStatus:
      type: string
      enum: [pending, cloning, configuring, ready, syncing, error, disconnected]
...
    RepoWorkspace:
      type: object
      additionalProperties: false
      required:
        [
          workspace,
          remote_branch,
          local_branch,
          head_commit_sha,
          tree_sha,
          status,
          updated_at,
        ]
      properties:
        workspace:
          $ref: "#/components/schemas/RepoWorkspaceKind"
        remote_branch:
          type: string
          minLength: 1
          maxLength: 255
        local_branch:
          type: string
          enum: [mmdash/code, mmdash/article, mmdash/result]
        head_commit_sha:
          type: [string, "null"]
          pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$"
        tree_sha:
          type: [string, "null"]
          pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$"
        status:
          $ref: "#/components/schemas/RepoWorkspaceStatus"
        updated_at:
          type: string
          format: date-time
...
    Repository:
      type: object
      additionalProperties: false
      required:
        [
          repository_id,
          project_id,
          provider,
          display_name,
          remote_url,
          default_branch,
          status,
          settings_version,
          last_synced_at,
          last_error_code,
          last_error_message,
          workspaces,
          webhook,
          created_at,
          updated_at,
        ]
      properties:
        repository_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        provider:
          $ref: "#/components/schemas/RepoProvider"
        display_name:
          type: string
        remote_url:
          description: Canonical GitHub URL. Local provider paths are always returned as null.
          type: [string, "null"]
        default_branch:
          type: string
        status:
          $ref: "#/components/schemas/RepositoryStatus"
        settings_version:
          type: integer
          minimum: 1
        last_synced_at:
          type: [string, "null"]
          format: date-time
        last_error_code:
          type: [string, "null"]
        last_error_message:
          type: [string, "null"]
        workspaces:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/RepoWorkspace"
        webhook:
          $ref: "#/components/schemas/RepoWebhook"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ModelOverview:
      type: object
      additionalProperties: false
      required: [project_id, generated_at, configured, discovered_pages, questions]
      properties:
        project_id: { type: string, format: uuid }
        generated_at: { type: string, format: date-time }
        configured: { type: boolean }
        source: { $ref: "#/components/schemas/ModelSource" }
        discovered_pages:
          type: array
          items: { $ref: "#/components/schemas/ModelSourcePage" }
        questions:
          type: array
          items: { $ref: "#/components/schemas/ModelQuestion" }
    ModelSource:
      type: object
      additionalProperties: false
      required:
        [source_id, project_id, notion_root_page_id, notion_root_page_url, notion_root_title, auto_sync_enabled, auto_sync_interval_seconds, sync_status, discovered_page_count, created_at, updated_at]
      properties:
        source_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        notion_root_page_id: { type: string, format: uuid }
        notion_root_page_url: { type: string, format: uri }
        notion_root_title: { type: string, maxLength: 255 }
        auto_sync_enabled: { type: boolean }
        auto_sync_interval_seconds: { type: integer, minimum: 60, maximum: 86400 }
        next_sync_at: { type: string, format: date-time }
        countdown_seconds: { type: integer, minimum: 0 }
        last_sync_id: { type: string, format: uuid }
        last_sync_status: { $ref: "#/components/schemas/ModelSyncStatus" }
        last_synced_at: { type: string, format: date-time }
        last_error_code: { type: string, maxLength: 200 }
        last_error_message: { type: string, maxLength: 2000 }
        sync_status: { $ref: "#/components/schemas/ModelSyncStatus" }
        discovered_page_count: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ModelQuestion:
      type: object
      additionalProperties: false
      required:
        [question_id, project_id, code, title, notion_page_id, notion_page_url, position, snapshot_count, sync_status, created_by, updated_by, created_at, updated_at]
      properties:
        question_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        code: { type: string, pattern: "^[A-Za-z][A-Za-z0-9_-]{0,31}$" }
        title: { type: string, minLength: 1, maxLength: 255 }
        notion_page_id: { type: string, format: uuid }
        notion_page_url: { type: string, format: uri }
        position: { type: integer, minimum: 0 }
        latest_snapshot_id: { type: string, format: uuid }
        snapshot_count: { type: integer, minimum: 0 }
        sync_status: { $ref: "#/components/schemas/ModelSyncStatus" }
        last_sync_id: { type: string, format: uuid }
        last_synced_at: { type: string, format: date-time }
        last_error_code: { type: string, maxLength: 200 }
        last_error_message: { type: string, maxLength: 2000 }
        created_by: { type: string, format: uuid }
        updated_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ModelQuestionList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ModelQuestion" }
    ModelQuestionDetail:
      type: object
      additionalProperties: false
      required: [question, snapshots]
      properties:
        question: { $ref: "#/components/schemas/ModelQuestion" }
        latest_snapshot: { $ref: "#/components/schemas/ModelSnapshot" }
        snapshots:
          type: array
          items: { $ref: "#/components/schemas/ModelSnapshotSummary" }
    CreateModelQuestionRequest:
      type: object
      additionalProperties: false
      required: [code, title, notion_page_id]
      properties:
        code: { type: string, pattern: "^[A-Za-z][A-Za-z0-9_-]{0,31}$" }
        title: { type: string, minLength: 1, maxLength: 255 }
        notion_page_id: { type: string, format: uuid }
        position: { type: integer, minimum: 0 }
    UpdateModelQuestionRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        code: { type: string, pattern: "^[A-Za-z][A-Za-z0-9_-]{0,31}$" }
        title: { type: string, minLength: 1, maxLength: 255 }
        notion_page_id: { type: string, format: uuid }
        position: { type: integer, minimum: 0 }
...
    ModelSync:
      type: object
      additionalProperties: false
      required: [sync_id, project_id, scope, trigger, status, job_id, requested_by, requested_at, updated_at]
      properties:
        sync_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        question_id: { type: string, format: uuid }
        scope: { type: string, enum: [source, question] }
        trigger: { type: string, enum: [manual, scheduled, settings] }
        status: { $ref: "#/components/schemas/ModelSyncStatus" }
        job_id: { type: string, format: uuid }
        requested_by: { type: string, format: uuid }
        requested_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        created_snapshot_id: { type: string, format: uuid }
        error_code: { type: string, maxLength: 200 }
        error_message: { type: string, maxLength: 2000 }
        updated_at: { type: string, format: date-time }
...
    ModelSnapshot:
      allOf:
        - $ref: "#/components/schemas/ModelSnapshotSummary"
        - type: object
          additionalProperties: false
          required: [project_id, notion_page_id, notion_page_url, outline, blocks, content_markdown, content_text, assets]
          properties:
            project_id: { type: string, format: uuid }
            notion_page_id: { type: string, format: uuid }
            notion_page_url: { type: string, format: uri }
            outline:
              type: array
              items: { $ref: "#/components/schemas/ModelOutlineItem" }
            blocks:
              type: array
              items: { $ref: "#/components/schemas/ModelBlock" }
            content_markdown: { type: string }
            content_text: { type: string }
            assets:
              type: array
              items: { $ref: "#/components/schemas/ModelSnapshotAsset" }
    ModelSnapshotList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ModelSnapshotSummary" }
    UpdateModelSnapshotRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 64 }
        version_note: { type: string, maxLength: 4000 }
...
    ModelDiff:
      type: object
      additionalProperties: false
      required: [question_id, from_snapshot_id, to_snapshot_id, granularity, blocks]
      properties:
        question_id: { type: string, format: uuid }
        from_snapshot_id: { type: string, format: uuid }
        to_snapshot_id: { type: string, format: uuid }
        granularity: { type: string, enum: [character] }
        blocks:
          type: array
          items: { $ref: "#/components/schemas/ModelDiffBlock" }
...
security:
  - bearerAuth: []